requirements.txt
scipy
pandas
matplotlib
ipython
python-decouple
maskpass==0.3.6
requests
casadi==3.5.5
pyyaml
joblib
cmake
seaborn